refactoring: readability of importing report
